pyproject.toml
[tool.poetry]
name = "cashu"
version = "0.6.0"
description = "Ecash wallet and mint."
authors = ["calle <callebtc@protonmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = "^3.7"
requests = "2.27.1"
pytest-asyncio = "0.19.0"
SQLAlchemy = "1.3.24"
click = "8.0.4"
pydantic = "^1.10.2"
bech32 = "^1.2.0"
fastapi = "^0.83.0"
environs = "^9.5.0"
uvicorn = "^0.18.3"
loguru = "^0.6.0"
ecdsa = "^0.18.0"
bitstring = "^3.1.9"
secp256k1 = "^0.14.0"
sqlalchemy-aio = "^0.17.0"
python-bitcoinlib = "^0.11.2"
h11 = "0.12.0"
PySocks = "^1.7.1"
cryptography = "^38.0.4"
websocket-client = "1.3.3"
pycryptodomex = "^3.16.0"

[tool.poetry.dev-dependencies]
black = {version = "^22.8.0", allow-prereleases = true}
isort = "^5.10.1"

[tool.poetry.group.dev.dependencies]
mypy = "^0.971"
black = {version = "^22.8.0", allow-prereleases = true}
isort = "^5.10.1"
pytest-cov = "^4.0.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.scripts]
mint = "cashu.mint.main:main"
cashu = "cashu.wallet.cli:cli"
wallet-test = "tests.test_wallet:test"